+++ Assets/Scripts/World/Chest.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 450f071b0514a2a4b9cc3df138da49f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
