+++ Assets/Scripts/World/Chest/InvisibilityChest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f01869637f9acf747aeb979dbea53e86
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
